pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "carcara"
version = "26.4.3"
authors = [
  { name="Leandro Seixas Rocha", email="leandro.rocha@ilum.cnpem.br" },
]
description = "Advancing materials research through autonomous on-the-fly active learning frameworks"
readme = "README.md"
license = {file = "LICENSE"} 
requires-python = ">=3.10"
dependencies = [
    'numpy>=2.0.0',
    'pandas',
    'ase',
    'torch',
    'scipy',
    'matplotlib',
    'pytest',
    'pyyaml',
    'ml_dtypes',
    'mace-torch'
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Physics"
]

[project.urls]
"Homepage" = "https://github.com/seixasgroup/carcara"
# "Documentation" = "https://seixasgroup.github.io/carcara/"
"Bug Tracker" = "https://github.com/seixasgroup/carcara/issues"
